Added type of frequency clarification
In reading the documentation for `fftfreqs`, I did not find what I came there for - a clarification about if the frequencies were measured in oscillations/second, or radians/second. From my background (engineering student), I am used to talking about angular frequencies in most cases, and so the fact that normal frequency is returned was not obvious to me, indication that it might not be to everyone. I have therefore appended the following line in the docs for `fftfreq` and `rfftfreq`:
`The return values are not to be confused with angular frequency.`

I considered specifying the unit of oscillation/second, but I felt like specifying this without adding "as opposed to..." was weird, and when adding it, it is longer and more trailing than this solution. Other suggestions are very welcome