For conversion of scale vector in adjoint
It always defines an `Array` which can fail on the GPU. This forces it to be the right type. One could also use `adapt` here, but since the element type promotion would have to occur anyways in the subsequent broadcast it seems you might as well convert all at once.